Drop url parsing logic
First of several cleanups. We used to support specifying VCFs and BAMs with
attributes, like '/path/to/file.bam#filter="I" in cigarstring'. That can be
somewhat convenient (especially when dealing with multple files with different
filters) but also awkward to explain and error prone. Got rid of that logic.
Now everything is in individual arguments.